Add artifact deletion error handling and only-metadata flag

+++ src/zenml/client.py
@@ -2288,31 +2288,50 @@ class Client(metaclass=ClientMetaClass):
         """
         return self.zen_store.get_artifact(artifact_id)
 
-    def delete_artifact(self, artifact_id: UUID) -> None:
+    def delete_artifact(
+        self, artifact_id: UUID, only_metadata: bool = False
+    ) -> None:
         """Delete an artifact.
 
-        This will delete both the metadata of the artifact from the database
-        and the actual artifact from the artifact store.
+        By default, this will delete both the metadata of the artifact from the
+        database and the actual artifact from the artifact store.
 
         Args:
             artifact_id: The ID of the artifact to delete.
+            only_metadata: If True, only delete the metadata of the artifact
+                and not the artifact itself.
         """
         from zenml.artifact_stores.base_artifact_store import BaseArtifactStore
         from zenml.stack.stack_component import StackComponent
 
         artifact = self.get_artifact(artifact_id=artifact_id)
-        if artifact.artifact_store_id:
-            artifact_store_model = self.get_stack_component(
-                component_type=StackComponentType.ARTIFACT_STORE,
-                name_id_or_prefix=artifact.artifact_store_id,
-            )
-            artifact_store = StackComponent.from_model(artifact_store_model)
-            assert isinstance(artifact_store, BaseArtifactStore)
-            artifact_store.rmtree(artifact.uri)
-            logger.info(
-                f"Deleted artifact '{artifact.uri}' from the artifact store."
-            )
 
+        # Delete the artifact from the artifact store.
+        if not only_metadata and artifact.artifact_store_id:
+            try:
+                artifact_store_model = self.get_stack_component(
+                    component_type=StackComponentType.ARTIFACT_STORE,
+                    name_id_or_prefix=artifact.artifact_store_id,
+                )
+                artifact_store = StackComponent.from_model(artifact_store_model)
+                assert isinstance(artifact_store, BaseArtifactStore)
+                artifact_store.rmtree(artifact.uri)
+            except Exception as e:
+                logger.error(
+                    f"Failed to delete artifact '{artifact.uri}' from the "
+                    "artifact store. This might happen if your local client "
+                    "does not have access to the artifact store or does not "
+                    "have the required integrations installed. Full error: "
+                    f"{e}"
+                )
+                raise e
+            else:
+                logger.info(
+                    f"Deleted artifact '{artifact.uri}' from the artifact "
+                    "store."
+                )
+
+        # Delete the artifact metadata from the database.
         self.zen_store.delete_artifact(artifact_id)
         logger.info(f"Deleted metadata of artifact '{artifact.uri}'.")
